refactor: validation and styles on the login page
