+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import GlobalStyles from './styles/global'
 
-import SignIn from './pages/SignUp'
+import SignIn from './pages/SignIn'
 
 const App: React.FC = () => (
   <>
