+++ src/pages/SignIn/index.tsx
@@ -1,5 +1,10 @@
-import React from 'react'
+import React, { useCallback, useRef } from 'react'
 import { FiLogIn, FiMail, FiLock } from 'react-icons/fi'
+import { Form } from '@unform/web'
+import { FormHandles } from '@unform/core'
+import * as Yup from 'yup'
+
+import getValidationErrors from '../../utils/getValidationErrors'
 
 import logoImg from '../../assets/logo.svg'
 
@@ -8,33 +13,59 @@ import Button from '../../components/Button'
 
 import { Container, Content, Background } from './styles'
 
-const SignIn: React.FC = () => (
-  <Container>
-    <Content>
-      <img src={logoImg} alt="GoBarber"/>
+const SignIn: React.FC = () => {
+  const formRef = useRef<FormHandles>(null)    
 
-      <form>
-        <h1>Faça seu logon</h1>
-
-        <Input icon={FiMail} name="email" placeholder="Email"/>
-
-        <Input icon={FiLock} name="password" type="password" placeholder="Senha"/>
-
-        <Button type="submit">Entrar</Button>
-
-        <a href="forgot">Esqueci minha senha</a>
-
-      </form>
+  const handleSubmit = useCallback(async (data: object) => {
+    try {
+      formRef.current?.setErrors({})
       
-        <a href="login">
-          <FiLogIn />
-          Criar conta
-        </a>
-    </Content>
+      const schema = Yup.object().shape({
+        email: Yup.string()
+          .required('E-mail obrigatório')
+          .email('Digite um e-mail válido'),
+        password: Yup.string().required('Senha obrigatória')
+      })
 
-    <Background />
-  </Container>
-)
+      await schema.validate(data, {
+        abortEarly: false
+      })
+
+    } catch (error) {
+      const errors = getValidationErrors(error)
+
+      formRef.current?.setErrors(errors)
+    }
+  }, [])
+
+  return (
+    <Container>
+      <Content>
+        <img src={logoImg} alt="GoBarber"/>
+
+        <Form ref={formRef} onSubmit={handleSubmit}>
+          <h1>Faça seu logon</h1>
+
+          <Input icon={FiMail} name="email" placeholder="Email"/>
+
+          <Input icon={FiLock} name="password" type="password" placeholder="Senha"/>
+
+          <Button type="submit">Entrar</Button>
+
+          <a href="forgot">Esqueci minha senha</a>
+
+        </Form>
+        
+          <a href="login">
+            <FiLogIn />
+            Criar conta
+          </a>
+      </Content>
+
+      <Background />
+    </Container>
+  )
+}
 
 export default SignIn
